ARTEMIS-3850 Eliminate Page Soft Cache. Add Option to read messages based on sizing.

+++ artemis-core-client/src/main/java/org/apache/activemq/artemis/api/core/management/AddressControl.java
@@ -98,7 +98,7 @@ public interface AddressControl {
     * Returns the number of pages used by this address.
     */
    @Attribute(desc = NUMBER_OF_PAGES_DESCRIPTION)
-   int getNumberOfPages();
+   long getNumberOfPages();
 
    /**
     * Returns whether this address is paging.
